refactor: MyActivityDisabled component 분리
